requirements-test.txt
nose
django_nose
ddt
